+++ src/package/index.ts
@@ -1,2 +1,4 @@
-export * from './DynamicForm';
-export * from './DynamicButton';
+export * from './DynamicForm'; // 表单
+export * from './DynamicButton'; // 按钮
+export * from './DynamicDropdown'; // 下拉框
+export * from './DynamicButtonGroup'; // 按钮组